refactor(fun): co-locate gambling service and view
