+++ bot/features/fun/gambling/__init__.py
@@ -1,0 +1,9 @@
+"""Interactive gambling workflow and Discord UI."""
+
+from .service import GamblingMixin
+from .view import GambleView
+
+__all__ = [
+    "GambleView",
+    "GamblingMixin",
+]
